Refactor dataset ingestion responsibilities
